+++ History.md
@@ -4,6 +4,9 @@
 * Fixed bug where an empty `fields` object was sometimes passed to a `changed`
   callback of `Cursor.observeChanges`.
 
+* Fixed `{$type: 5}` selectors for binary values on browsers that do not support
+  `Uint8Array`
+
 ## v0.5.7
 
 * The DDP wire protocol has been redesigned.